+++ test/drivers/usbled.js
@@ -1,68 +1,86 @@
-var buster = require('buster-node'),
-  fs = require('fs'),
-  referee = require('referee'),
-  UsbLed = require('../../lib/drivers/usbled'),
-  assert = referee.assert;
+"use strict"
+/* eslint no-unused-vars: 0 */
+import fs from 'fs';
+import UsbLed from '../../lib/drivers/usbled.js';
+import referee from '@sinonjs/referee';
+import sinon from 'sinon';
+const assert = referee.assert;
 
-buster.testCase('usbled - usbled', {
-  setUp: function () {
-    this.mockFs = this.mock(fs);
-  },
-  'should set path': function () {
-    var usbled = new UsbLed('/some/path');
+describe('usbled - usbled', function() {
+  beforeEach(function () {
+    this.mockFs = sinon.mock(fs);
+  });
+  afterEach(function () {
+    this.mockFs.verify();
+    this.mockFs.restore();
+  });
+  it('should set path', function () {
+    const usbled = new UsbLed('/some/path');
     assert.equals(usbled.path, '/some/path');
-  }
+  });
 });
 
-buster.testCase('usbled - on', {
-  setUp: function () {
-    this.mockFs = this.mock(fs);
-  },
-  'should write 1 to colour file': function () {
+describe('usbled - on', function() {
+  beforeEach(function () {
+    this.mockFs = sinon.mock(fs);
+  });
+  afterEach(function () {
+    this.mockFs.verify();
+    this.mockFs.restore();
+  });
+  it('should write 1 to colour file', function () {
     this.mockFs.expects('writeFileSync').once().withExactArgs('/some/path/red', 1);
     new UsbLed('/some/path').on('red'); 
-  },
-  'should handle case insensitive colours': function () {
+  });
+  it('should handle case insensitive colours', function () {
     this.mockFs.expects('writeFileSync').twice().withExactArgs('/some/path/red', 1);
-    var usbled = new UsbLed('/some/path');
+    const usbled = new UsbLed('/some/path');
     usbled.on('RED');
     usbled.on('rEd');
-  }
+  });
 });
 
-buster.testCase('usbled - off', {
-  setUp: function () {
-    this.mockFs = this.mock(fs);
-  },
-  'should write 0 to colour file': function () {
+describe('usbled - off', function() {
+  beforeEach(function () {
+    this.mockFs = sinon.mock(fs);
+  });
+  afterEach(function () {
+    this.mockFs.verify();
+    this.mockFs.restore();
+  });
+  it('should write 0 to colour file', function () {
     this.mockFs.expects('writeFileSync').once().withExactArgs('/some/path/red', 0);
     new UsbLed('/some/path').off('red'); 
-  },
-  'should handle case insensitive colours': function () {
+  });
+  it('should handle case insensitive colours', function () {
     this.mockFs.expects('writeFileSync').twice().withExactArgs('/some/path/red', 0);
-    var usbled = new UsbLed('/some/path');
+    const usbled = new UsbLed('/some/path');
     usbled.off('RED');
     usbled.off('rEd');
-  }
+  });
 });
 
-buster.testCase('usbled - _find', {
-  setUp: function () {
-    this.mockFs = this.mock(fs);
-  },
-  'should return path when there is only one version of usbled installed': function () {
+describe('usbled - _find', function() {
+  beforeEach(function () {
+    this.mockFs = sinon.mock(fs);
+  });
+  afterEach(function () {
+    this.mockFs.verify();
+    this.mockFs.restore();
+  });
+  it('should return path when there is only one version of usbled installed', function () {
     this.mockFs.expects('existsSync').once().withExactArgs('/sys/bus/usb/drivers/usbled/').returns(true);
     this.mockFs.expects('readdirSync').once().withExactArgs('/sys/bus/usb/drivers/usbled/').returns(['2-1.8:1.0']);    
     this.usbLed = new UsbLed();
     assert.equals(this.usbLed.path, '/sys/bus/usb/drivers/usbled/2-1.8:1.0');
-  },
-  'should return path with newer version when there are multiple versions of usbled installed': function () {
+  });
+  it('should return path with newer version when there are multiple versions of usbled installed', function () {
     this.mockFs.expects('existsSync').once().withExactArgs('/sys/bus/usb/drivers/usbled/').returns(true);
     this.mockFs.expects('readdirSync').once().withExactArgs('/sys/bus/usb/drivers/usbled/').returns(['2-1.8:1.0', 'bind', '2-0.1:2.3']);
     this.usbLed = new UsbLed();
     assert.equals(this.usbLed.path, '/sys/bus/usb/drivers/usbled/2-1.8:1.0');
-  },
-  'should throw an error when usbled exists but there is no version': function (done) {
+  });
+  it('should throw an error when usbled exists but there is no version', function (done) {
     this.mockFs.expects('existsSync').once().withExactArgs('/sys/bus/usb/drivers/usbled/').returns(true);
     this.mockFs.expects('readdirSync').once().withExactArgs('/sys/bus/usb/drivers/usbled/').returns([]);
     try {
@@ -71,8 +89,8 @@ buster.testCase('usbled - _find', {
       assert.equals(e.message, 'Unable to find USB LED driver installation.');
       done();
     }
-  },
-  'should throw an error when there is no usbled version installed': function (done) {
+  });
+  it('should throw an error when there is no usbled version installed', function (done) {
     this.mockFs.expects('existsSync').once().withExactArgs('/sys/bus/usb/drivers/usbled/').returns(true);
     this.mockFs.expects('readdirSync').once().withExactArgs('/sys/bus/usb/drivers/usbled/').returns(['bind', 'uevent', 'unbind']);
     try {
@@ -81,8 +99,8 @@ buster.testCase('usbled - _find', {
       assert.equals(e.message, 'Unable to find USB LED driver installation.');
       done();
     }
-  },
-  'should throw an error when usbled base directory does not exist': function (done) {
+  });
+  it('should throw an error when usbled base directory does not exist', function (done) {
     this.mockFs.expects('existsSync').once().withExactArgs('/sys/bus/usb/drivers/usbled/').returns(false);
     try {
       new UsbLed();
@@ -90,5 +108,5 @@ buster.testCase('usbled - _find', {
       assert.equals(e.message, 'Unable to find USB LED driver installation.');
       done();
     }
-  }
+  });
 });